Fix remote existence detection
There was a condition where if the remote wasn't responding or had
errors, we would think it wasn't installed on the system, and then fail
the install (because the remote was having issues). Since we were
basically matching on name only anyway, this does the same thing, but
doesn't fail if the remote is erroring.

There should still be a better way to do this, but for now, this is
better than the previous method.